requirements.txt
alembic==1.7.7
amqp==5.1.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
Authlib==1.0.1
Babel==2.11.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
browser-cookie3==0.16.1
bs4==0.0.1
celery==5.2.6
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
cryptography==38.0.4
cssselect==1.1.0
dateparser==1.1.1
defusedxml==0.7.1
demjson3==3.0.5
Deprecated==1.2.13
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==8.5.2
email-validator==1.1.3
facebook-scraper==0.2.59
fake-useragent==0.1.11
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httpie==3.2.1
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.0.1
jeepney==0.8.0
Jinja2==3.0.1
joblib==1.1.0
keyring==23.7.0
kombu==5.2.4
lxml==4.9.1
lz4==4.0.1
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
ndg-httpsclient==0.5.1
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.1
parse==1.19.0
pbkdf2==1.3
prometheus-client==0.14.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pusher==3.3.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pyee==8.2.2
Pygments==2.12.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
redis==4.2.2
regex==2022.3.2
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rq==1.11.1
scrapingbee==1.1.8
SecretStorage==3.3.2
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.32
stripe==2.68.0
textblob==0.17.1
tornado==6.1
tqdm==4.63.1
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.8.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.0.3
wrapt==1.14.0
wsproto==1.2.0
WTForms==3.0.1
zipp==3.8.0
